chore: separated card component out, removed test case
