+++ packages/p2p/src/components/modal-manager/modals/share-my-ads-modal/share-my-ads-card/index.ts
@@ -1,0 +1,4 @@
+import ShareMyAdsCard from './share-my-ads-card';
+import './share-my-ads-card.scss';
+
+export default ShareMyAdsCard;
