+++ packages/p2p/src/components/modal-manager/modals/share-my-ads-modal/share-my-ads-card/share-my-ads-card.scss
@@ -1,0 +1,76 @@
+.share-my-ads-card {
+    background: linear-gradient(349deg, var(--general-section-1) 36%, var(--brand-red-coral) 25%);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    margin-bottom: 1.5rem;
+    padding: 2rem;
+    position: relative;
+
+    @include mobile {
+        width: 90%;
+    }
+
+    &__icon {
+        height: 16px;
+        width: max-content;
+        margin: 1rem 0 3rem;
+
+        @include mobile {
+            margin-top: 0;
+        }
+    }
+
+    &__numbers {
+        display: flex;
+        flex-direction: row;
+
+        &-text {
+            display: flex;
+            flex-direction: column;
+            margin-bottom: 1rem;
+
+            &:first-child {
+                margin-right: 1rem;
+            }
+
+            &:last-child {
+                margin-bottom: 0;
+            }
+        }
+    }
+
+    &__qr {
+        align-items: center;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        margin-top: 2rem;
+
+        @include mobile {
+            margin-top: 1rem;
+        }
+
+        &-container {
+            background-color: #fff;
+            display: flex;
+            padding: 0.4rem;
+            border: 0.1rem solid var(--general-active);
+            border-radius: 1rem;
+            position: relative;
+            width: fit-content;
+        }
+
+        &-text {
+            margin: 1.5rem 0 2rem;
+
+            @include mobile {
+                margin-bottom: 0;
+            }
+        }
+    }
+
+    &__title {
+        margin-bottom: 2rem;
+    }
+}
