+++ packages/p2p/src/components/modal-manager/modals/share-my-ads-modal/share-my-ads-modal.scss
@@ -45,86 +45,9 @@
                 }
             }
 
-            &-details {
-                background: linear-gradient(349deg, var(--general-section-1) 36%, var(--brand-red-coral) 25%);
-                display: flex;
-                flex-direction: column;
-                justify-content: center;
-                margin-bottom: 1.5rem;
-                padding: 2rem;
-                position: relative;
-
-                @include mobile {
-                    width: 90%;
-                }
-
-                &-icon {
-                    height: 16px;
-                    width: max-content;
-                    margin: 1rem 0 3rem;
-
-                    @include mobile {
-                        margin-top: 0;
-                    }
-                }
-
-                &--numbers {
-                    display: flex;
-                    flex-direction: row;
-
-                    &-text {
-                        display: flex;
-                        flex-direction: column;
-                        margin-bottom: 1rem;
-
-                        &:first-child {
-                            margin-right: 1rem;
-                        }
-
-                        &:last-child {
-                            margin-bottom: 0;
-                        }
-                    }
-                }
-
-                &-title {
-                    margin-bottom: 2rem;
-                }
-            }
-
             &__download-button {
                 @include mobile {
                     width: 100%;
-                }
-            }
-
-            &-qr {
-                align-items: center;
-                display: flex;
-                flex-direction: column;
-                justify-content: center;
-                margin-top: 2rem;
-
-                @include mobile {
-                    margin-top: 1rem;
-                }
-
-                &__background {
-                    background-color: #fff;
-                    display: flex;
-                    padding: 0.4rem;
-                    border: 0.1rem solid var(--general-active);
-                    border-radius: 1rem;
-                    position: relative;
-                    width: fit-content;
-                }
-
-                &__text {
-                    margin: 1.5rem 0 2rem;
-
-                    @include mobile {
-                        margin-bottom: 0;
-                    }
                 }
             }
         }
