--- packages/p2p/src/components/modal-manager/modals/share-my-ads-modal/share-my-ads-socials/__tests__/share-my-ads-icons.spec.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { screen, render } from '@testing-library/react';
-import ShareMyAdsSocials from '../share-my-ads-socials';
-
-describe('<ShareMyAdsSocials />', () => {
-    it('should render all the social media platforms', () => {
-        render(<ShareMyAdsSocials />);
-
-        expect(screen.getByText('WhatsApp')).toBeInTheDocument();
-        expect(screen.getByText('Facebook')).toBeInTheDocument();
-        expect(screen.getByText('Telegram')).toBeInTheDocument();
-        expect(screen.getByText('Twitter')).toBeInTheDocument();
-        expect(screen.getByText('Gmail')).toBeInTheDocument();
-    });
-});
